+++ src/os2datascanner/projects/admin/adminapp/migrations/0117_move_rules_to_rule_fields.py
@@ -8,20 +8,37 @@ def set_rule(apps, schema_editor):
     Scanner = apps.get_model('os2datascanner', 'Scanner')
 
     for scanner in Scanner.objects.iterator():
-        print("\nScanner pk:", scanner.pk)
         if scanner.rules.count() > 1:
-            print(f"Scanner '{scanner}' has more than one connected rule! Creating a new rule.")
-            date_stamp = time_now().strftime('%y%m%d')
+            # The scanner has multiple rules. Create a new rule that combines all
+            # existing rules.
+            date_stamp = time_now().strftime('%y-%m-%d')
             rules = scanner.rules.all()
             new_rule, _ = CustomRule.objects.get_or_create(
                 name=f"{date_stamp} - Migration rule ({', '.join([str(rule.pk) for rule in rules])})", 
-                description=f"OrRule that combines the rules: {', '.join([f'{rule.name} ({rule.pk})' for rule in rules])}",
-                _rule={"type": "or", "compontents": [rule.customrule._rule for rule in rules]}
-            )
+                defaults={
+                "description": f"OrRule that combines the rules: {', '.join([f'{rule.name} ({rule.pk})' for rule in rules])}",
+                "_rule": {"type": "or", "components": [rule.customrule._rule for rule in rules]},
+                "organization": scanner.organization
+                }
+                )
             scanner.rule = new_rule
             scanner.save()
-        else:
+        elif scanner.rules.count() == 1:
+            # The scanner has one rule. Use that.
             scanner.rule = scanner.rules.first()
+            scanner.save()
+        else:
+            # The scanner has no rules. Create a dummy rule and use that.
+            date_stamp = time_now().strftime('%y-%m-%d')
+            dummy_rule, _ = CustomRule.objects.get_or_create(
+                name=f"{date_stamp} - Dummy rule", 
+                defaults={
+                    "description": "Dummy rule created as a part of migration to new rule field, since no rules could be succesfully moved from old rules field.",
+                    "_rule": {},
+                    "organization": scanner.organization
+                    }
+                )
+            scanner.rule = dummy_rule
             scanner.save()
 
 def set_exclusion_rule(apps, schema_editor):
@@ -30,8 +47,9 @@ def set_exclusion_rule(apps, schema_editor):
 
     for scanner in Scanner.objects.iterator():
         if scanner.exclusion_rules.count() > 1:
-            print(f"Scanner '{scanner}' has more than one connected rule! Creating a new rule.")
-            date_stamp = time_now().strftime('%y%m%d')
+            # The scanner has multiple exclusion rules. Create a new rule that 
+            # combines all existing rules.
+            date_stamp = time_now().strftime('%y-%m-%d')
             rules = scanner.exclusion_rules.all()
             new_rule, _ = CustomRule.objects.get_or_create(
                 name=f"{date_stamp} - Migration rule ({', '.join([str(rule.pk) for rule in rules])})", 
@@ -40,12 +58,16 @@ def set_exclusion_rule(apps, schema_editor):
                     "_rule": {"type": "or", "components": [rule.customrule._rule for rule in rules]},
                     "organization": scanner.organization
                     }
-            )
+                )
             scanner.exclusion_rule = new_rule
             scanner.save()
-        else:
+        elif scanner.exclusion_rules.count() == 1:
+            # The scanner has one exclusion rule. Use that.
             scanner.exclusion_rule = scanner.exclusion_rules.first()
             scanner.save()
+        else:
+            # The scanner has no exclusion rules. Don't add any.
+            pass
 
 def revert_rule_to_rules(apps, schema_editor):
     Scanner = apps.get_model('os2datascanner', 'Scanner')