+++ src/os2datascanner/projects/admin/adminapp/migrations/0118_remove_scanner_rules_and_exclusion_rules_fields.py
@@ -9,6 +9,11 @@ class Migration(migrations.Migration):
     ]
     
     operations = [
+        migrations.AlterField(
+            model_name='scanner',
+            name='rule',
+            field=models.ForeignKey(null=False, on_delete=django.db.models.deletion.PROTECT, related_name='scanners', to='os2datascanner.rule', verbose_name='rule'),
+        ),
         migrations.RemoveField(
             model_name='scanner',
             name='exclusion_rules',