+++ src/os2datascanner/projects/admin/adminapp/views/dropboxscanner_views.py
@@ -35,13 +35,13 @@ class DropboxScannerCreate(ScannerCreate):
     fields = [
         'name',
         'schedule',
-        'exclusion_rules',
+        'exclusion_rule',
         'token',
         'do_ocr',
         'do_last_modified_check',
         'keep_false_positives',
         'only_notify_superadmin',
-        'rules',
+        'rule',
         'organization',
     ]
 
@@ -66,13 +66,13 @@ class DropboxScannerUpdate(ScannerUpdate):
     fields = [
         'name',
         'schedule',
-        'exclusion_rules',
+        'exclusion_rule',
         'token',
         'do_ocr',
         'do_last_modified_check',
         'keep_false_positives',
         'only_notify_superadmin',
-        'rules',
+        'rule',
         'organization',
     ]