+++ src/os2datascanner/projects/admin/tests/test_utilities.py
@@ -1,0 +1,5 @@
+dummy_rule_dict = {
+    "name": "dummy rule",
+    "description": "this is a dumb dumb dummy rule",
+    "_rule": {"type": "regex", "expression": "dummy"}
+}